requirements.txt
langchain
langchain-community
langchain-core
langchain-litellm
python-dotenv
pydantic
dspy-ai
